layout: TeslaLaborTimes - finish
